Modif w/Redux

+++ API/controllers/tasks/taskActions.js
@@ -13,8 +13,12 @@ const GetTasksAction = (TaskModel) => async (req, res) => {
 const GetTaskByIdAction = (TaskModel) => async (req, res) => {
     try {
         const { id } = req.params;
-        const task = await TaskModel.findById(id);
-        return res.status(200).json(task);
+        if (id) {
+            const task = await TaskModel.findById(id);
+            return res.status(200).json(task);
+        } else {
+            res.formatter.ok({ msg: "No id in the request" });
+        }
     } catch (error) {
         return res.formatter.ok({ msg: "Task not found" });
     }
@@ -23,8 +27,12 @@ const GetTaskByIdAction = (TaskModel) => async (req, res) => {
 const DeleteTaskAction = (TaskModel) => async (req, res) => {
     try {
         const { id } = req.params;
-        const deletedTasks = await TaskModel.findByIdAndDelete(id);
-        return res.status(200).json(deletedTasks);
+        if (id) {
+            const deletedTasks = await TaskModel.findByIdAndDelete(id);
+            return res.status(200).json(deletedTasks);
+        } else {
+            res.formatter.ok({ msg: "No id in the request" });
+        }
     } catch (error) {
         return res.formatter.ok({ msg: "Can't delete task" });
     }
@@ -43,9 +51,13 @@ const CreateTaskAction = (TaskModel) => async (req, res) => {
 const UpdateTaskAction = (TaskModel) => async (req, res) => {
     try {
         const { id } = req.params;
-        await TaskModel.updateOne({ _id: id }, req.body);
-        const updatedTasks = await TaskModel.findById(id);
-        return res.status(200).json(updatedTasks);
+        if (id && req.body) {
+            await TaskModel.updateOne({ _id: id }, req.body);
+            const updatedTasks = await TaskModel.findById(id);
+            return res.status(200).json(updatedTasks);
+        } else {
+            return res.formatter.ok({ msg: "No id / body in the request" });
+        }
     } catch (error) {
         return res.formatter.ok({ msg: "Can't update task" });
     }
